requirements.txt
# Core dependencies
pandas>=2.2.0
openpyxl>=3.1.0
pandasql>=0.7.3
openai>=1.0.0
python-dotenv>=1.0.0

# UI Framework
streamlit>=1.30.0

# Utilities
sqlparse>=0.5.0

# Optional: For enhanced features
# guardrails-ai>=0.5.10  # Uncomment for production guardrails
